SocialShare: add no-referral generation into the mix
In preparation for 2947 to enable the Social Buttons for Unlisted Claims, we will need a version of the short URL without any referral code (in addition to one with the code, as the main share link).

It wasn't clear from the history why a no-referral version was used for the Social Buttons, but there are sub-reddits that ban links with referral codes, so the existing behavior was retained.